переносит рендер пагинации во внешний модуль
